Replace 'result delay' with 'publication date'
Selecting a 'result delay' option is very limiting for admins; this
switches it to a datetime picker allowing them to set an arbitrary time
to post results. On the back end, not much changes: we still track
a result delay by computing the difference in seconds between the end
and publish times.